requirements.txt
aiogram==2.14.3
python-binance==1.0.16
numpy==1.21.2
pandas==1.3.3
python-dotenv==0.19.0
requests==2.31.0
aiohttp==3.8.1
TA-Lib==0.4.24
python-telegram-bot==20.7
cryptography==41.0.7
stripe==2.60.0
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
aioredis==2.0.1
python-jose==3.3.0
jinja2==3.0.1
python-multipart==0.0.5
redis==3.5.3
psycopg2-binary==2.9.1 